+++ test/helpers/Fixture.ts
@@ -1,7 +1,7 @@
 import { ethers, hubbleBls } from "../../deps/index.ts";
 
 import Rng from "./Rng.ts";
-import contractABIs from "../../contractABIs/index.ts";
+import ovmContractABIs from "../../ovmContractABIs/index.ts";
 import createBLSWallet from "./createBLSWallet.ts";
 import WalletService from "../../src/app/WalletService.ts";
 import { TransactionData } from "../../src/app/TxService.ts";
@@ -104,7 +104,7 @@ export default class Fixture {
   connectBlsWallet(address: string) {
     return new ethers.Contract(
       address,
-      contractABIs["BLSWallet.json"].abi,
+      ovmContractABIs["BLSWallet.json"].abi,
       this.walletService.aggregatorSigner,
     );
   }
